Update application settings and improve email domain block process
Updated application settings to get the host from the environment variable and added a signal to automate the email domain blocking process after saving a client. Also, a dependency `docutils` has been added to `requirements.txt`, a method in `utils.py` has been modified to continue the process even if the status code is not 200, and the client saving process and success message have been simplified in `app/views.py`.