Renamed 'client_id' to 'client_key' and added Mastodon integration
In this update, 'client_id' field in client model has been renamed to 'client_key'. Also added integration with Mastodon API, which allows the application to register with the Mastodon instance, generate an authorization URL, obtain an access token using the authorization code, and send a signal once the client is authorized. Included is a new function for handling the retrieval and storage of access tokens. A migration file for this change is also provided. Additionally, the FAQ page and its associated URL mapping have been introduced.