Refactor cron jobs to asynchronous tasks
The purpose of this commit was to improve application efficiency by shifting from using cron jobs to utilizing asynchronous tasks managed by Django-Q. The redundant use of Django-cron has been removed, and the 'test.py' file has been deleted. Furthermore, the implementation of rate limiting ensures that the application does not exceed the API's call limit. Finally, this commit includes code enhancements to several files and the addition of new dependencies to the requirements.txt file.