Add fixture data and set FAQ text rendering as safe
This commit adds a db.json file under app/fixtures for setting up fixture data for FAQs. These FAQs are added to the setup process in the Docker compose file. Furthermore, the rendering of FAQ text in the faq.html template file is now set as safe to prevent character issues.