Remove cronjob, update app, and Docker adjustments
The commit removes the app-cronjob.sh file previously used for running cronjobs, and instead adjusts the Django application to handle these operations. A backup schedule is added in the Django schedule update, and supervisor is added to the Dockerfile setup. Adjustments are made across files to improve the Docker operation, including changes to .dockerignore, Dockerfile, and docker-compose.yml.